requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-simple-history==3.5.0
djangorestframework==3.14.0
python-dateutil==2.6.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
